backend/python/requirements.txt
numpy
pandas
torch
fastapi[standard]
python-dotenv
